backend/requirements.txt
fastapi
uvicorn
jinja2
httpx
requests
python-multipart  # For form handling
python-jose[cryptography]  # JWT authentication
passlib[bcrypt]  # Password hashing
python-dotenv  # Environment variables
supabase  # Database integration
redis  # Caching layer
pydantic  # Data validation
sqlalchemy  # ORM
alembic  # Database migrations
reportlab  # PDF generation
weasyprint  # HTML to PDF conversion
aiofiles  # Async file operations
python-multipart  # Form handling
